Perform WHOIS queries
WHOIS queries get stored in the ./whois-data directory (which is created
when you start the server if it doesn't exist). At startup, the WHOIS
manager waits 1 minute before doing its first check for domains with the
'enabled' property. If they do not have a file in ./whois-data it will do
a lookup and save a file.

If there is a file for that domain, it reads that file and checks

- If it has been 9 months since it was updated
- If the domain expires in 3 months (+- 24 hours)
- If the domain expires in 2 months (+- 24 hours)
- If the domain expires in 1 month (+- 24 hours)
- If the domain expires in 2 weeks (+- 24 hours)

If any of those conditions are true it will perform a query and update
the whois data stored for that domain.

The whois manager then waits 24 hours before starting the process over again.

Plan to add a way to force a whois query for a domain.