refactor: :boom: source code restructuring; using templating
Breaking up monolithic index.html into partials and templating them together, makes it easier to work on. Same option is available for the javascript files now too. Using 11ty to do this (with a little fs.appendFileSync magic). Restructured the code into server and client folders, and the dockerfile now copies 3 times into the /app. Maybe this can be a little neater in the future.

Testing indicates that everything behaves the exact same way as it did before.